Refactor cases controller into 2 separate controllers.
This will split the actions between those requiring a user (portfolio actions)
and the previously existing `create` action, referring to GLiMR.
