+++ app/views/users/cases/_case_row.html.erb
@@ -12,8 +12,8 @@
   <% end %>
   <td class="right">
     <span class="right actions actions-tight">
-      <%= button_to t('.delete'), case_path(tribunal_case), method: :delete, data: {confirm: t('.delete_confirmation')}, class: 'button button-secondary' %>
-      <%= link_to t('.resume'), case_resume_path(tribunal_case), class: 'button' %>
+      <%= button_to t('.delete'), users_case_path(tribunal_case), method: :delete, data: {confirm: t('.delete_confirmation')}, class: 'button button-secondary' %>
+      <%= link_to t('.resume'), users_case_resume_path(tribunal_case), class: 'button' %>
     </span>
   </td>
 </tr>
